Revert "sources: Allow to install debuginfo packages"
This reverts commit c2a88176a4cd414f9dd85d54bf5ff03126b26486.

Previously, debuginfo component was set in the Docker image used
for tests. This led to installation of debuginfo RPM packages even
it was not needed.

For example,
```
apt-get install .*ipa-client

Reading Package Lists...
Building Dependency Tree...
Selecting freeipa-client-samba for '.*ipa-client'
Selecting freeipa-client-automount for '.*ipa-client'
Selecting freeipa-client for '.*ipa-client'
Selecting freeipa-client-common for '.*ipa-client'
Selecting freeipa-client-debuginfo for '.*ipa-client'
Selecting freeipa-client-epn for '.*ipa-client'
```

So apt debuginfo component is only added if it's really needed for
installation of debuginfo packages to get readable tracebacks.